+++ src/stdio/printf/pf_handle_string.c
@@ -1,20 +1,8 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   pf_handle_string.c                                 :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: nalysann <urbilya@gmail.com>               +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2020/08/27 10:56:42 by nalysann          #+#    #+#             */
-/*   Updated: 2020/08/27 10:56:43 by nalysann         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
+#include <stdarg.h>
 
 #include "pf_handle_placeholder.h"
 
 #include "ft_string.h"
-
-#include <stdarg.h>
 
 char	*handle_string(t_fields *fields, va_list ap)
 {
@@ -24,7 +12,7 @@ char	*handle_string(t_fields *fields, va_list ap)
 	str = va_arg(ap, char *);
 	if (!str)
 		str = "(null)";
-	len = ft_strlen(str);
+	len = (int)ft_strlen(str);
 	if (fields->precision != GET_DEFAULT && len > fields->precision)
 		len = fields->precision;
 	return (ft_strndup(str, len));
