+++ src/stdio/printf/pf_print_placeholder.c
@@ -1,26 +1,14 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   pf_print_placeholder.c                             :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: nalysann <urbilya@gmail.com>               +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2020/08/27 10:57:17 by nalysann          #+#    #+#             */
-/*   Updated: 2020/08/27 10:57:18 by nalysann         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
+#include <stdarg.h>
+#include <stdbool.h>
+#include <stdlib.h>
+
+#include "ft_string.h"
 
 #include "pf_buffer.h"
 #include "pf_handle_placeholder.h"
 #include "pf_type_handlers.h"
 
-#include "ft_string.h"
-
-#include <stdarg.h>
-#include <stdbool.h>
-#include <stdlib.h>
-
-static const t_type_handler		g_type_handlers[TYPES_SIZE + 1] =
+static const t_type_handler	g_type_handlers[TYPES_SIZE + 1] =
 {
 	handle_char,
 	handle_string,
@@ -37,7 +25,7 @@ static const t_type_handler		g_type_handlers[TYPES_SIZE + 1] =
 	handle_invalid
 };
 
-static bool		is_zero(char *str)
+static bool	is_zero(char *str)
 {
 	while (*str)
 		if (*str++ != '0')
@@ -45,7 +33,7 @@ static bool		is_zero(char *str)
 	return (true);
 }
 
-static char		*handle_prefix(t_fields *fields, char *result)
+static char	*handle_prefix(t_fields *fields, char *result)
 {
 	if (fields->type != 'c' && *result == '-')
 	{
@@ -56,9 +44,9 @@ static char		*handle_prefix(t_fields *fields, char *result)
 		return ("+");
 	if ((fields->flags & FLAG_SPACE))
 		return (" ");
-	if ((fields->flags & FLAG_HASH) && fields->type == 'o' &&
-		((ft_strcmp(result, "0") && fields->precision <= 0) ||
-		(result[0] != '0')))
+	if ((fields->flags & FLAG_HASH) && fields->type == 'o'
+		&& ((ft_strcmp(result, "0") && fields->precision <= 0)
+			|| (result[0] != '0')))
 		return ("0");
 	if ((fields->flags & FLAG_HASH) && fields->type == 'x' && !is_zero(result))
 		return ("0x");
@@ -69,14 +57,14 @@ static char		*handle_prefix(t_fields *fields, char *result)
 	return ("");
 }
 
-static int		print_result(t_fields *fields, char *result, t_buffer *buf)
+static int	print_result(t_fields *fields, char *result, t_buffer *buf)
 {
 	char	*prefix;
 	int		len;
 	int		i;
 
 	prefix = handle_prefix(fields, result);
-	len = ft_strlen(prefix) + ft_strlen(result);
+	len = (int)ft_strlen(prefix) + (int)ft_strlen(result);
 	if (fields->type == 'c' && !*result)
 		++len;
 	if (ft_strlen(prefix) != 0 && (fields->flags & FLAG_ZERO) != 0)
@@ -96,17 +84,17 @@ static int		print_result(t_fields *fields, char *result, t_buffer *buf)
 	return ((fields->width > len) ? fields->width : len);
 }
 
-int				print_placeholder(t_fields *fields, va_list ap, t_buffer *buf)
+int	print_placeholder(t_fields *fields, va_list ap, t_buffer *buf)
 {
 	int		ret;
 	char	*result;
 
-	result = g_type_handlers[ft_strchr(SUPPORTED_TYPES, fields->type) -
-								SUPPORTED_TYPES](fields, ap);
+	result = g_type_handlers[ft_strchr(SUPPORTED_TYPES, fields->type)
+		- SUPPORTED_TYPES](fields, ap);
 	if (!result)
 		return (-1);
-	if (fields->type == 'f' &&
-		(ft_strstr(result, "inf") || ft_strstr(result, "nan")))
+	if (fields->type == 'f'
+		&& (ft_strstr(result, "inf") || ft_strstr(result, "nan")))
 		fields->flags &= ~FLAG_ZERO;
 	if (fields->type == 'f' && ft_strstr(result, "nan"))
 		fields->flags &= ~FLAG_PLUS & ~FLAG_SPACE;
